Remove MissingMappingsEvent
This dated event had an implementation which seemed overly complex for the job.
In its place is the IForgeRegistry/DeferredRegister addAlias API, which allows redirecting an entry with one name to another.
A better API can be thought out at a later date if it becomes apparent that modders still rely on functionality this event provided that cannot be replicated with the addAlias system.